context sensative help on Linux

+++ src/utils/launch.ts
@@ -12,6 +12,7 @@ export const launch = (target: string, online: boolean) => {
   if (online) shell.openExternal(target);
   else if (os.platform() === 'win32') shell.openItem(target);
   else {
+    console.log(target)
     const cmd = /\.sh/i.test(target) ? '' : 'xdg-open ';
     execa.command(`${cmd}${target}`, {
       env: { ...{ ...process }.env, DISPLAY: ':0' },
@@ -21,11 +22,21 @@ export const launch = (target: string, online: boolean) => {
 
 export const launchCmd = (target: string) => {
   if (!temp) throw new Error('Unable to find temp directory.'); //this is app.getPath('temp')
-  const tempName = path.join(temp, 'transcriber-cmd.ps1');
-  fs.writeFileSync(tempName, target, { encoding: 'utf-8' });
-  execa(`powershell`, [tempName]).finally(() => {
-    fs.unlinkSync(tempName);
-  });
+  if (os.platform() === 'win32') {
+    const tempName = path.join(temp, 'transcriber-cmd.ps1');
+    fs.writeFileSync(tempName, target, { encoding: 'utf-8' });
+    execa(`powershell`, [tempName]).finally(() => {
+      fs.unlinkSync(tempName);
+    });
+    } else {
+      const tempName = path.join(temp, 'transcriber-cmd.sh');
+      fs.writeFileSync(tempName, target, {encoding: 'utf-8'});
+      execa(`sh`, [tempName], {
+        env: { ...{ ...process }.env, DISPLAY: ':0' },
+      }).finally(() => {
+        fs.unlinkSync(tempName);
+      })
+    }
 };
 
 export default launch;
